% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTLComb.R
\name{CV_multi_outcomes}
\alias{CV_multi_outcomes}
\title{k-fold CV for multiple outcomes}
\usage{
CV_multi_outcomes(
  X,
  Y,
  nfolds = 10,
  nlambda = 200,
  lam_ratio = 0.001,
  ctasks = 4,
  opts = NULL,
  C = 1,
  C2 = 0.1,
  intercept = F
)
}
\arguments{
\item{X}{A set of feature matrices (the mixed tasks with classification and regression types)}

\item{Y}{A set of responses, could be binary (classification problem) or continues (regression problem). The valid
value of binary outcome \eqn{\in\{1, -1\}}}

\item{nfolds}{The number of folds in CV. The default value is 20}

\item{nlambda}{The number of lambda in the sequence. The default value is 10}

\item{lam_ratio}{The number (between 0 and 1) to determine the smallest lambda according to lam_max. For example 
lam_ratio * lam_max. The default value is 0.01}

\item{ctasks}{The index of the classification tasks. For example \code{ctasks=c(2,3)} refers that 2th and 3th tasks 
are classification}

\item{opts}{Options of the optimization procedure. One can set the initial search point, the tolerance and the 
maximized number of iterations and the termination rule using this parameter. The default value is 
\code{list(init=0,  tol=10^-4, maxIter=50, ter=2)}}

\item{C}{A non-negative constant to promote the selection of same signatures}

\item{C2}{A non-negative constant to improve the generalization performance. The default value is 0}

\item{intercept}{True/False indicator to tell whether intercept or intercept-free model}

\item{lambda}{A positive constant \eqn{\lambda} to control the number of jointly selected features}
}
\value{
The CV results
}
\description{
Split the subjects of each task into k folds, training a MTLComb model on k-1 folds of data and test on the holding out fold. 
Repeat this procedure over all folds
}
\author{
Dr. Han Cao
}
