% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTLComb.R
\name{getCVPartition}
\alias{getCVPartition}
\title{Get the data splits for CV}
\usage{
getCVPartition(Y, cv_fold)
}
\arguments{
\item{Y}{A set of responses, could be binary (classification problem) or continues (regression problem). The valid
value of binary outcome \eqn{\in\{1, -1\}}}

\item{cv_fold}{The number of folds for CV}
}
\value{
The data splits of all folds
}
\description{
The subjects of each task will be randomly divided into k folds
}
\author{
Dr. Han Cao
}
