% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTLComb.R
\name{MTLComb_CV}
\alias{MTLComb_CV}
\title{k-fold CV}
\usage{
MTLComb_CV(
  X = NULL,
  Y = NULL,
  nfolds = 10,
  lam_ratio = 0.01,
  nlambda = 10,
  lambda = NULL,
  ctasks = NULL,
  intercept = F,
  opts = list(init = 0, maxIter = 50, tol = 0.001, ter = 2),
  C = 0
)
}
\arguments{
\item{X}{A set of feature matrices (the mixed tasks with classification and regression types)}

\item{Y}{A set of responses, could be binary (classification problem) or continues (regression problem). The valid
value of binary outcome \eqn{\in\{1, -1\}}}

\item{nfolds}{The number of folds in CV. The default value is 20}

\item{lam_ratio}{The number (between 0 and 1) to determine the smallest lambda according to lam_max. For example 
lam_ratio * lam_max. The default value is 0.01}

\item{nlambda}{The number of lambda in the sequence. The default value is 10}

\item{lambda}{A positive constant \eqn{\lambda} to control the number of jointly selected features}

\item{ctasks}{The index of the classification tasks. For example \code{ctasks=c(2,3)} refers that 2th and 3th tasks 
are classification}

\item{intercept}{True/False indicator to tell whether intercept or intercept-free model}

\item{opts}{Options of the optimization procedure. One can set the initial search point, the tolerance and the 
maximized number of iterations and the termination rule using this parameter. The default value is 
\code{list(init=0,  tol=10^-4, maxIter=50, ter=2)}}

\item{C}{A non-negative constant to improve the generalization performance. The default value is 0}
}
\value{
The CV results
}
\description{
Split the subjects of each task into k folds, training a MTLComb model on k-1 folds of data and test on the holding out fold. 
Repeat this procedure over all folds
}
