% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTLComb.R
\name{MTLComb_Train}
\alias{MTLComb_Train}
\title{Regularization path estimation}
\usage{
MTLComb_Train(
  X = NULL,
  Y = NULL,
  nlambda = 10,
  lam_ratio = 0.01,
  lambda = NULL,
  C = 0,
  C2 = 0,
  ctasks = NULL,
  intercept = F,
  opts = list(init = 0, maxIter = 20, tol = 0.01, ter = 2)
)
}
\arguments{
\item{X}{A set of feature matrices (the mixed tasks with classification and regression types)}

\item{Y}{A set of responses, could be binary (classification problem) or continues (regression problem). The valid
value of binary outcome \eqn{\in\{1, -1\}}}

\item{nlambda}{The number of lambda in the sequence. The default value is 10}

\item{lam_ratio}{The number (between 0 and 1) to determine the smallest lambda according to lam_max. For example 
lam_ratio * lam_max. The default value is 0.01}

\item{lambda}{A positive constant \eqn{\lambda} to control the number of jointly selected features}

\item{C}{A non-negative constant to promote the selection of same signatures}

\item{C2}{A non-negative constant to improve the generalization performance. The default value is 0}

\item{ctasks}{The index of the classification tasks. For example \code{ctasks=c(2,3)} refers that 2th and 3th tasks 
are classification}

\item{intercept}{True/False to indicator whether to use intercept or intercept-free model}

\item{opts}{Options of the optimization procedure. One can set the initial search point, the tolerance and the 
maximized number of iterations and the termination rule using this parameter. The default value is 
\code{list(init=0,  tol=10^-2, maxIter=100, ter=2)}}
}
\value{
The regularization path
}
\description{
Train a series of MTLComb models indexed by a sequence of lambda
}
\author{
Dr. Han Cao
}
