% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTLComb.R
\name{plot_MTLComb_CV}
\alias{plot_MTLComb_CV}
\title{Plot the CV result}
\usage{
plot_MTLComb_CV(cvResult)
}
\arguments{
\item{cvResult}{The result from calling the function \code{MTLComb_CV}}
}
\description{
Plot the CV result for both regression and clasification tasks
}
