% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTLComb.R
\name{predict_MTLComb}
\alias{predict_MTLComb}
\title{Prediction with MTLComb model}
\usage{
predict_MTLComb(fit, newx, type)
}
\arguments{
\item{fit}{A MTLComb model}

\item{newx}{A new feature matrix}

\item{type}{True/False indicator to show whether the new dataset is regression or classification}
}
\value{
The prediction scores of all tasks of \code{type} tasks
}
\description{
All regression or classification models would be predicted in the single targeting dataset. The prediction scores 
of all models are returned
}
\author{
Dr. Han Cao
}
