% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTLComb.R
\name{solve_L21}
\alias{solve_L21}
\title{The solver of MTLComb}
\usage{
solve_L21(X, Y, lam, C, C2, ctasks, opts)
}
\arguments{
\item{X}{A set of feature matrices (the mixed tasks with classification and regression types)}

\item{Y}{A set of responses, could be binary (classification problem) or continues (regression problem). The valid
value of binary outcome \eqn{\in\{1, -1\}}}

\item{lam}{A positive constant \eqn{\lambda} to control the number of jointly selected features}

\item{C}{A non-negative constant to promote the selection of same signatures}

\item{C2}{A non-negative constant to improve the generalization performance}

\item{ctasks}{the index of the classification tasks. For example \code{ctasks=c(2,3)} refered to 2th and 3th tasks 
are classification}

\item{opts}{Options of the optimization procedure. One can set the initial search point, the tolerance and the 
maximized number of iterations and the termination rule using this parameter. For example 
\code{list(init=0,  tol=10^-3, maxIter=1000, ter=2)}}
}
\value{
The coefficients of the model and the historical objective values
}
\description{
Solve the objective of MTLComb
}
\author{
Dr. Han Cao
}
